backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
Pillow==10.4.0
pytesseract==0.3.13
python-Levenshtein==0.26.0
pydantic==2.9.2
